Fixed add attachments automation
