+++ frontend/src/tools/AddAttachments.tsx
@@ -6,10 +6,8 @@ import { BaseToolProps, ToolComponent } from "../types/tool";
 import { useEndpointEnabled } from "../hooks/useEndpointConfig";
 import { useAddAttachmentsParameters } from "../hooks/tools/addAttachments/useAddAttachmentsParameters";
 import { useAddAttachmentsOperation } from "../hooks/tools/addAttachments/useAddAttachmentsOperation";
-import { Stack, Text, Group, ActionIcon, Alert, ScrollArea, Button } from "@mantine/core";
-import LocalIcon from "../components/shared/LocalIcon";
 import { useAccordionSteps } from "../hooks/tools/shared/useAccordionSteps";
-// Removed FitText for two-line wrapping with clamping
+import AddAttachmentsSettings from "../components/tools/addAttachments/AddAttachmentsSettings";
 
 const AddAttachments = ({ onPreviewFile, onComplete, onError }: BaseToolProps) => {
   const { t } = useTranslation();
@@ -67,99 +65,11 @@ const AddAttachments = ({ onPreviewFile, onComplete, onError }: BaseToolProps) =
       onCollapsedClick: () => accordion.handleStepToggle(AddAttachmentsStep.ATTACHMENTS),
       isVisible: true,
       content: (
-        <Stack gap="md">
-          <Alert color="blue" variant="light">
-            <Text size="sm">
-              {t("AddAttachmentsRequest.info", "Select files to attach to your PDF. These files will be embedded and accessible through the PDF's attachment panel.")}
-            </Text>
-          </Alert>
-
-          <Stack gap="xs">
-            <Text size="sm" fw={500}>
-              {t("AddAttachmentsRequest.selectFiles", "Select Files to Attach")}
-            </Text>
-            <input
-              type="file"
-              multiple
-              onChange={(e) => {
-                const files = Array.from(e.target.files || []);
-                // Append to existing attachments instead of replacing
-                const newAttachments = [...params.parameters.attachments, ...files];
-                params.updateParameter('attachments', newAttachments);
-                // Reset the input so the same file can be selected again
-                e.target.value = '';
-              }}
-              disabled={endpointLoading}
-              style={{ display: 'none' }}
-              id="attachments-input"
-            />
-            <Button
-              size="xs"
-              color="blue"
-              component="label"
-              htmlFor="attachments-input"
-              disabled={endpointLoading}
-              leftSection={<LocalIcon icon="plus" width="14" height="14" />}
-            >
-              {params.parameters.attachments.length > 0 
-                ? t("AddAttachmentsRequest.addMoreFiles", "Add more files...")
-                : t("AddAttachmentsRequest.placeholder", "Choose files...")
-              }
-            </Button>
-          </Stack>
-
-          {params.parameters.attachments && params.parameters.attachments.length > 0 && (
-            <Stack gap="xs">
-              <Text size="sm" fw={500}>
-                {t("AddAttachmentsRequest.selectedFiles", "Selected Files")} ({params.parameters.attachments.length})
-              </Text>
-              <ScrollArea.Autosize mah={300} type="scroll" offsetScrollbars styles={{ viewport: { overflowX: 'hidden' } }}>
-                <Stack gap="xs">
-                  {params.parameters.attachments.map((file, index) => (
-                    <Group key={index} justify="space-between" p="xs" style={{ border: '1px solid var(--mantine-color-gray-3)', borderRadius: 'var(--mantine-radius-sm)', alignItems: 'flex-start' }}>
-                      <Group gap="xs" style={{ flex: 1, minWidth: 0, alignItems: 'flex-start' }}>
-                        {/* Filename (two-line clamp, wraps, no icon on the left) */}
-                        <div style={{ flex: 1, minWidth: 0 }}>
-                          <div
-                            style={{
-                              fontSize: 'var(--mantine-font-size-sm)',
-                              fontWeight: 400,
-                              lineHeight: 1.2,
-                              display: '-webkit-box',
-                              WebkitLineClamp: 2 as any,
-                              WebkitBoxOrient: 'vertical' as any,
-                              overflow: 'hidden',
-                              whiteSpace: 'normal',
-                              wordBreak: 'break-word',
-                            }}
-                            title={file.name}
-                          >
-                            {file.name}
-                          </div>
-                        </div>
-                        <Text size="xs" c="dimmed" style={{ flexShrink: 0 }}>
-                          ({(file.size / 1024).toFixed(1)} KB)
-                        </Text>
-                      </Group>
-                      <ActionIcon
-                        size="sm"
-                        variant="subtle"
-                        color="red"
-                        style={{ flexShrink: 0 }}
-                        onClick={() => {
-                          const newAttachments = params.parameters.attachments.filter((_, i) => i !== index);
-                          params.updateParameter('attachments', newAttachments);
-                        }}
-                      >
-                        <LocalIcon icon="close-rounded" width="14" height="14" />
-                      </ActionIcon>
-                    </Group>
-                  ))}
-                </Stack>
-              </ScrollArea.Autosize>
-            </Stack>
-          )}
-        </Stack>
+        <AddAttachmentsSettings
+          parameters={params.parameters}
+          onParameterChange={params.updateParameter}
+          disabled={endpointLoading}
+        />
       ),
     });
